% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_acoes.R
\name{buscar_num_acoes}
\alias{buscar_num_acoes}
\title{CRAWLER NUMERO DE ACOES EMPRESA}
\usage{
buscar_num_acoes(tickers)
}
\arguments{
\item{tickers}{texto com o ticker da empresa}
}
\value{
retorna o numero de acoes da empresa
}
\description{
Realiza um crawler no site fundamentus para capturar o numero de acoes que uma empresa
listada em bolsa possui
}
\examples{
buscar_num_acoes("PETR3")
}
\author{
Pedro Cavalcante
}
