% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP.R
\name{construir_bp_ativo}
\alias{construir_bp_ativo}
\title{CONSTRUIR BALANCO PATRIMONIAL ATIVO}
\usage{
construir_bp_ativo(bulk)
}
\arguments{
\item{bulk}{lista com os dados brutos}
}
\value{
balanco patrimonial ativo
}
\description{
Constroi o balanço patrimonial ativo apos os dados terem sido devidamente capturados
}
\author{
Pedro Cavalcante
}
