% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP.R
\name{construir_bp_passivo}
\alias{construir_bp_passivo}
\title{CONSTRUIR BALANCO PATRIMONIAL PASSIVO}
\usage{
construir_bp_passivo(bulk)
}
\arguments{
\item{bulk}{lista com os dados brutos}
}
\value{
balanco patrimonial passivo
}
\description{
Constroi o balanço patrimonial passivo apos os dados terem sido devidamente capturados
}
\author{
Pedro Cavalcante
}
