% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker_empresa.R
\name{construir_depara}
\alias{construir_depara}
\alias{rm_accent}
\title{CONSTRUIR DEPARA}
\usage{
rm_accent(str, pattern = "all")
}
\arguments{
\item{bulk}{lista com os dados brutos}
}
\value{
lista com a tabela de tickers disponiveis e o nome das empresas
}
\description{
Realiza o depara entre diferentes fontes de dados (B3, CVM e Fundamentus)
}
\author{
Pedro Cavalcante
}
