% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRE.R
\name{construir_DRE}
\alias{construir_DRE}
\title{CONSTRUIR DRE}
\usage{
construir_DRE(bulk)
}
\arguments{
\item{bulk}{lista com os dados brutos}
}
\value{
demonstrativo de resultados
}
\description{
Constroi o demonstrativo de resultados apos os dados terem sido devidamente capturados
}
\author{
Pedro Cavalcante
}
