% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construir_metrica.R
\name{construir_metricas}
\alias{construir_metricas}
\title{Criar Metricas}
\usage{
construir_metricas(tbl_dre, tbl_bp, empresas_selec)
}
\arguments{
\item{tbl_dre}{Tabela de Demonstrativo de Resultados}

\item{tbl_bp}{Tabela de Balanco Patrimonial}

\item{empresas_selec}{Selecionar Empresas para construir metricas}
}
\value{
retorna df com as informacoes das metricas
}
\description{
Constroi metricas a partir de tbl_dre, tbl_bp
}
\author{
Pedro Cavalcante
}
