% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_bd.R
\name{importar_bd}
\alias{importar_bd}
\title{Importa banco de dados}
\usage{
importar_bd(diretorio)
}
\arguments{
\item{diretorio}{Local em que as tabelas serao geradas em csv}
}
\value{
Retorna uma lista com as tabelas
}
\description{
Funcao principal que importa o banco de dados
}
\details{
Funcao principal que importa o banco de dados,
}
\examples{
importar_bd("")


}
\author{
Pedro Cavalcante
}
