% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGS.R, R/sidra.R
\name{importar_sidra}
\alias{importar_sidra}
\alias{importar_sgs}
\title{API SIDRA}
\usage{
importar_sgs(info)

importar_sidra(info)
}
\arguments{
\item{info}{Variavel de busca que pode ser "inflacao","pib","abate","PIM","PMS_VOLUME","PMS_RECEITA"#'}
}
\value{
dataframe com a data e com a varavel requisitada

dataframe com a data e com a varavel requisitada
}
\description{
Variavel que se utiliza da API do ipeadatar para capturar alguns dados relevantes à analise

Variavel que se utiliza da API do sidra para capturar alguns dados relevantes à analise
}
\examples{
importar_sgs("juros")
importar_sidra("inflacao")
}
\author{
Pedro Cavalcante

Pedro Cavalcante
}
