% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B3.R, R/fundamentus.R
\name{ticker_empresa}
\alias{ticker_empresa}
\alias{buscar_setores}
\title{DEPARA SETOR NOME EMPRESA}
\usage{
buscar_setores()

ticker_empresa()
}
\value{
dataframe com nome da empresa e seus setores

dataframe com nome do ticker, nome da companhia e nome que esta na razao social
}
\description{
A partir de um crawler oriundo do site da b3, cria um depara entre a empresa e setor atuante

A partir de um crawler oriundo do site fundamentus, cria um depara entre o ticker da empresa listada
na bolsa e a sau razao social
}
\examples{
buscar_setores()
}
\author{
Pedro Cavalcante

Pedro Cavalcante
}
